build(static_centerline_generator): prefix package and namespace with autoware_ (#6817)
* build(static_centerline_generator): prefix package and namespace with autoware_


* style(pre-commit): autofix


* build: fix CMake target


* build(autoware_static_centerline_generator): more renames


* style(pre-commit): autofix


* build(autoware_static_centerline_generator): fix namespace


* fix(autoware_static_centerline_generator): fix clang-tidy issues


* style(pre-commit): autofix


* style(pre-commit): autofix


* fix(autoware_static_centerline_generator): fix clang-tidy issues


* fix(autoware_static_centerline_generator): fix build issues


* fix(autoware_static_centerline_generator): fix build issues


* style(pre-commit): autofix


* fix(autoware_static_centerline_optimizer): fix clang-tidy issues


* style(pre-commit): autofix


* build: fix build errors


* fix: remove else statements after return


* fix(autoware_static_centerline_generator): fix clang-tidy issues


* style(pre-commit): autofix


* revert changes for static_centerline_generator


* fix(autoware_static_centerline_generator): add autoware_ prefix


* style(pre-commit): autofix


* fix(autoware_static_centerline_generator): fix filenames


* fix(autoware_static_centerline_generator): fix namespaces


* style(pre-commit): autofix


* fix: added prefix to missing strings


* refactor(autoware_static_centerline_generator): move header files to src


* refactor(autoware_static_centerline_generator): fix include paths


* style(pre-commit): autofix


* refactor(autoware_static_centerline_generator): rename base folder


* Update planning/autoware_static_centerline_generator/launch/static_centerline_generator.launch.xml


* build(autoware_static_centerline_generator): fix include in CMake


* build(autoware_static_centerline_generator): fix missing includes


---------
