+++ planning/autoware_static_centerline_generator/config/static_centerline_generator.param.yaml
