+++ planning/autoware_static_centerline_generator/config/vehicle_info.param.yaml
