+++ planning/autoware_static_centerline_generator/launch/run_planning_server.launch.xml
@@ -1,0 +1,12 @@
+<launch>
+  <!-- mandatory arguments for planning-->
+  <arg name="vehicle_model"/>
+
+  <include file="$(find-pkg-share autoware_static_centerline_generator)/launch/static_centerline_generator.launch.xml">
+    <arg name="vehicle_model" value="$(var vehicle_model)"/>
+    <arg name="run_background" value="true"/>
+  </include>
+
+  <!-- local server to connect path optimizer and cloud software -->
+  <node pkg="autoware_static_centerline_generator" exec="app.py" name="static_centerline_generator_http_server" output="screen"/>
+</launch>
