+++ planning/autoware_static_centerline_generator/media/rviz.png
