+++ planning/autoware_static_centerline_generator/media/unsafe_footprints.png
