+++ planning/autoware_static_centerline_generator/msg/PointsWithLaneId.msg
