+++ planning/autoware_static_centerline_generator/srv/PlanRoute.srv
