+++ planning/autoware_static_centerline_generator/test/data/bag_ego_trajectory.db3
