+++ planning/autoware_static_centerline_generator/test/data/lanelet2_map.osm
