--- planning/static_centerline_generator/scripts/tmp/bag_ego_trajectory_based_centerline.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-ros2 launch static_centerline_generator static_centerline_generator.launch.xml centerline_source:="bag_ego_trajectory_base" run_background:=false lanelet2_input_file_path:="$HOME/autoware_map/sample_map/lanelet2_map.osm" bag_filename:="$(ros2 pkg prefix static_centerline_generator --share)/test/data/bag_ego_trajectory.db3" vehicle_model:=lexus
