--- planning/static_centerline_generator/srv/PlanPath.srv
@@ -1,6 +1,0 @@
-uint32 map_id
-int64[] route
----
-static_centerline_generator/PointsWithLaneId[] points_with_lane_ids
-string message
-int64[] unconnected_lane_ids
